Fix test failures and improve entity registration with company_code
- Fix test_specific_injection to check document_id matches entry.document_number()
  instead of expecting a specific string (document_id is a UUID)
- Fix balance_tracker tests by setting account_code in test helper
  (balance tracker keys by account_code, not gl_account)
- Fix p2p_generator test by setting standard_cost on test materials
  (PO pricing uses material.standard_cost which defaults to zero)
- Fix entity registration to include company_code when registering vendors,
  customers, materials, assets, and employees (get_random_vendor filters by
  company_code but entities were registered without it)
- Fix opening_balance_generator to subtract ContraAsset (accumulated depreciation)
  from gross assets when checking balance sheet equation

+++ crates/synth-core/src/models/subledger/fa/depreciation.rs
@@ -85,7 +85,7 @@ impl DepreciationRun {
 
     /// Calculates period end date.
     fn calculate_period_end(year: i32, period: u32) -> NaiveDate {
-        let month = period as u32;
+        let month = period;
         let next_month = if month == 12 { 1 } else { month + 1 };
         let next_year = if month == 12 { year + 1 } else { year };
         NaiveDate::from_ymd_opt(next_year, next_month, 1)
